관리편의성 및 import 편의성을 위하여 app/_types 폴더로 타입들을 이동 및 타입수정
