+++ app/_types/chat.ts
@@ -1,13 +1,10 @@
-import { Socket } from 'socket.io-client';
-
 export type MessageType = 'chat' | 'donation' | 'recommend';
 export type TabType = 'chat' | 'viewers';
 export type Message = ChatMessage | DonationMessage | RecommendMessage;
+export type UserRole = 'broadcaster' | 'manager' | 'member' | 'viewer' | 'guest';
 
-export interface BaseMessage {
-  id?: string;
+interface BaseMessage {
   type: MessageType;
-  timestamp?: number;
 }
 
 export interface ChatMessage extends BaseMessage {
@@ -17,8 +14,8 @@ export interface ChatMessage extends BaseMessage {
   chatter_nickname: string;
   chatter_message: string;
   grade: string;
-  role: string;
   color: string;
+  jwt_decode: JwtDecode;
 }
 
 export interface DonationMessage extends BaseMessage {
@@ -33,16 +30,27 @@ export interface RecommendMessage extends BaseMessage {
   recommender_nickname: string;
 }
 
+export interface JwtDecode {
+  idx: number;
+  user_id: string;
+  nickname: string;
+  role: UserRole;
+  profile_img: string;
+  is_guest: boolean;
+  guest_id?: string;
+}
+
 export interface Viewer {
   user_idx: number;
   user_id: string;
   nickname: string;
-  type: 'viewer' | 'broadcaster' | 'manager';
+  role: JwtDecode;
 }
 
-export interface ChatProps {
-  broadcasterId: string;
-  socket: Socket | null;
-  myRole: 'broadcaster' | 'manager' | 'member' | 'guest';
-  broadcasterIdx: number;
+export interface MyRole {
+  user_idx: number;
+  user_id: string;
+  nickname: string;
+  role: UserRole
 }
+
